Update Snowball
I replaced the set called pattern ,because if not ,when add a pattern object to set ,it cannot be hashed so there is an error occured . Maybe there is a better way to solve this problem . But for now i can only change it to a list so there is problem of  hasing.  If im wrong please show me a better way. I would appreciate it . Thanks.